Adds client authentication
In order to interact with the Reddit API, a praw.Reddit object is
created with the appropriate client_id and client_secret (each new user
must create a new application). These are stored in an external file
which is passed as the first argument in the script.

An example file is also added to guide new users.